Organized and created containers and components, created upload button
