+++ client/containers/Header.jsx
@@ -1,0 +1,17 @@
+import React, { Component } from "react";
+
+export default class Header extends Component {
+    constructor(props) {
+        super(props);
+    }
+    
+    render () {
+
+        return (
+            <div className="header">
+                Header
+            </div>
+        )
+        
+    }
+}
