+++ client/index.html
@@ -5,6 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Sortera</title>
+  <script src="https://kit.fontawesome.com/a076d05399.js"></script>
 </head>
 <body>
